SymEx: Always require postcondition in verify* methods
To make the core method simpler we always require postcondition to be
given.
If there is no postcondition, we are not really "verifying".
However, there is a use case where we don't want to verify anything, we
just want to explore a contract.
For such use case, we provide a new helper method to do exactly that.
`exploreContract` executes the contract symbolically, and returns its
Expr representation.