Expr: Remove expensive unnecessary simplification
This transformation is a bit expensive to check and should not be
attempted repeatedly, which is what we are doing now.
It might make sense to apply it once at the very end before we encode to
SMT.
However, the simplification system does not support that yet.
For now the simplification can be removed.
We could attempt to bring it back later.