CEx: Fix finalizing INIT frame
When initialization code runs successfully, we still need to pay for
storing the deployed code.
This can turn successful frame into unsuccessful one, since we still may
run into out-of-gas.
When we are sure we have successfully deployed a contract, we need to
add its address to the known (warm) addresses.